// next-ragbot-app/interfaces/message.ts
export interface IMessage {
  user_type: UserType;
  message: string;
  send_date: Date;
}

export enum UserType {
  USER = "User",
  BOT = "Bot",
}
export enum BotType {
  LLMGPT = "LLMGPT",
  LLMGEMINI = "LLMGEMINI",
  RAGDOC = "RAGDOC",
  RAGWEB = "RAGWEB",
}

export interface ISendMessage {
  role: string;
  message: string;
}

//대화이력챗봇 전용 메시지 타입 정의: 기본메시지타입 상속받아 기능확장함
export interface IMemberMessage extends IMessage {
  nick_name: string;
}